+++ cmd/deploy.go
@@ -33,7 +33,7 @@ var Deploy = func(providers app.Providers) *cli.Command {
 		},
 		Action: func(c *cli.Context) error {
 			return AppWorkspaceAction(c, func(ctx context.Context, cfg api.Config, appDetails app.Details) error {
-				osWriters := logging.StandardOsWriters{}
+				osWriters := CliOsWriters{Context: c}
 				version, wait := c.String("version"), c.IsSet("wait")
 
 				commitSha := ""