+++ cmd/logs.go
@@ -3,7 +3,6 @@ package cmd
 import (
 	"context"
 	"github.com/nullstone-io/deployment-sdk/app"
-	"github.com/nullstone-io/deployment-sdk/logging"
 	"github.com/nullstone-io/deployment-sdk/outputs"
 	"github.com/urfave/cli/v2"
 	"gopkg.in/nullstone-io/go-api-client.v0"
@@ -79,8 +78,9 @@ var Logs = func(providers app.Providers) *cli.Command {
 			}
 
 			return AppWorkspaceAction(c, func(ctx context.Context, cfg api.Config, appDetails app.Details) error {
+				osWriters := CliOsWriters{Context: c}
 				source := outputs.ApiRetrieverSource{Config: cfg}
-				logStreamer, err := providers.FindLogStreamer(logging.StandardOsWriters{}, source, appDetails)
+				logStreamer, err := providers.FindLogStreamer(osWriters, source, appDetails)
 				if err != nil {
 					return err
 				}